enable test coverage for armv8-linux-musleabihf
 * fix libc prototypes of offsets to have correct integer sizes and
   signedness. This gets all behavior tests to pass for
   armv8-linux-musleabihf
 * fix linux mmap syscall to have correct integer size and signedness
   for offset
 * disable failing armv8-linux-musleabihf std lib tests. See 3289. I
   suspect compiler-rt issue. Note these tests fail with this target
   triple regardless of whether musl is actually linked (-lc).

+++ std/heap.zig
@@ -780,7 +780,7 @@ test "FixedBufferAllocator" {
 }
 
 test "FixedBufferAllocator.reset" {
-    var buf: [8]u8 align(@alignOf(usize)) = undefined;
+    var buf: [8]u8 align(@alignOf(u64)) = undefined;
     var fba = FixedBufferAllocator.init(buf[0..]);
 
     const X = 0xeeeeeeeeeeeeeeee;
